requirements.txt
asgiref==3.2.10
boto3==1.16.54
botocore==1.19.54
certifi==2020.12.5
chardet==4.0.0
Django==3.0.8
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.1
Pillow==8.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
s3transfer==0.3.4
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
